Restructure testing into separate files. Add API general tests.
